Splitting out HTML and SVG renderer code
