+++ src/render/dom/create-dom-visual-element.ts
@@ -1,7 +1,7 @@
 import { ComponentType } from "react"
 import { VisualElementOptions } from "../types"
-import { htmlVisualElement } from "./html-visual-element"
-import { svgVisualElement } from "./svg-visual-element"
+import { htmlVisualElement } from "../html/visual-element"
+import { svgVisualElement } from "../svg/visual-element"
 import { isSVGComponent } from "./utils/is-svg-component"
 
 export function createDomVisualElement(Component: string | ComponentType) {
