+++ src/render/dom/utils/unit-conversion.ts
@@ -3,7 +3,7 @@ import { Target, TargetWithKeyframes } from "../../../types"
 import { isKeyframesTarget } from "../../../animation/utils/is-keyframes-target"
 import { invariant } from "hey-listen"
 import { MotionValue } from "../../../value"
-import { transformProps } from "./transform"
+import { transformProps } from "../../html/utils/transform"
 import { AxisBox2D } from "../../../types/geometry"
 import { findDimensionValueType } from "./value-types"
 import { VisualElement } from "../../types"
