+++ src/render/svg/utils/camel-case-attrs.ts
