+++ src/render/svg/utils/create-render-state.ts
@@ -1,4 +1,4 @@
-import { createHtmlRenderState } from "./create-html-render-state"
+import { createHtmlRenderState } from "../../html/utils/create-render-state"
 import { zeroDimensions } from "./zero-dimensions"
 
 export const createSvgRenderState = () => ({
