+++ src/render/svg/utils/path.ts
