+++ src/render/svg/utils/zero-dimensions.ts
